+++ src/components/PDFViewer.tsx
@@ -5,6 +5,7 @@ import { PageViewMode, PageType } from '../types/Page';
 import { PDFViewerPage } from './PDFViewerPage';
 import { dataURItoUint8Array, isDataURI } from '../utils/hacks';
 import { PDFViewerToolbar, ToolbarLabelProps } from './PDFViewerToolbar';
+import { PDFViewerTouchToolbar } from './PDFViewerTouchToolbar';
 import { PDFWorker } from './PDFWorker';
 
 interface FullscreenableElement {
@@ -12,10 +13,14 @@ interface FullscreenableElement {
 }
 
 const DocumentWrapper = styled.div`
-  background-color: red;
+  background-color: #fff;
   position: relative;
   width: 100%;
-  padding-top: 56.25%; /* 16:9 Aspect Ratio */
+  height: calc(56vw + 40px); // Compensate for the area we remove in the document
+  max-height: 100vh;
+  // Thanks chrome for Android for not calculating the viewport size correctly depending
+  // on whether you show or not the address bar. But no worries, we'll do it manually
+  max-height: calc(var(--vh, 1vh) * 100);
   overflow: hidden;
 
   ${({ fullscreen }: FullscreenableElement) =>
@@ -26,6 +31,7 @@ const DocumentWrapper = styled.div`
       left: 0;
       right: 0;
       bottom: 0;
+      height: auto;
       position: fixed;
       z-index: 100500;
     `};
@@ -35,11 +41,12 @@ const Document = styled.div`
   position: absolute;
   top: 0;
   left: 0;
-  bottom: 0;
+  bottom: 40px;
   right: 0;
   background-color: ${themed(({ theme = {} }: StandardProps) => theme.ui2)};
   padding: ${distance.medium};
   overflow: scroll;
+  touch-action: pan-x pan-y;
 `;
 
 const PageWrapper = styled.div`
@@ -108,6 +115,19 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
   const [currentViewMode, setCurrentViewMode] = React.useState<PageViewMode>(PageViewMode.DEFAULT);
   const [currentScale, setCurrentScale] = React.useState(1);
   const [fullscreen, setFullscreen] = React.useState(false);
+  const [showTouchToolbar, setShowTouchToolbar] = React.useState(false);
+
+  const deviceAgent = navigator.userAgent.toLowerCase();
+
+  const isTouchDevice =
+    deviceAgent.match(/(iphone|ipod|ipad)/) ||
+    deviceAgent.match(/(android)/) ||
+    deviceAgent.match(/(iemobile)/) ||
+    deviceAgent.match(/iphone/i) ||
+    deviceAgent.match(/ipad/i) ||
+    deviceAgent.match(/ipod/i) ||
+    deviceAgent.match(/blackberry/i) ||
+    deviceAgent.match(/bada/i);
 
   /**
    * Every time a new file is set we load the new document
@@ -127,7 +147,14 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
    * Effect responsible for registering/unregistering the resize spy to determine the rendering sizes
    */
   React.useLayoutEffect(() => {
-    const handleResize = debounce(() => zoomToPageView(pages[currentPage], currentViewMode), 500);
+    const handleResize = debounce(() => {
+      zoomToPageView(pages[currentPage], currentViewMode);
+
+      // Fix chrome on Android address bar issue by setting the right viewport height with good old fashion JS
+      // Then we set the value in the --vh custom property to the root of the document
+      const vh = window.innerHeight * 0.01;
+      window.document.documentElement.style.setProperty('--vh', `${vh}px`);
+    }, 500);
     window.addEventListener('resize', handleResize);
 
     return () => window.removeEventListener('resize', handleResize);
@@ -210,12 +237,140 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
   }
 
   /**
+   * Touch events here
+   */
+  window['touchInfo'] = {
+    startX: 0,
+    startY: 0,
+    startDistance: 0,
+  };
+  const touchInfo = window['touchInfo'];
+
+  /**
+   * Event triggered when the user touches the screen
+   */
+  function onDocumentTouch() {
+    if (isTouchDevice) {
+      setShowTouchToolbar(!showTouchToolbar);
+    }
+  }
+
+  /**
+   * Event triggered on double touch
+   */
+  function onDocumentDoubleTouch() {
+    if (isTouchDevice) {
+      switchFullscreenMode();
+    }
+  }
+
+  let currentPinchScale = currentScale;
+
+  /**
+   * Event triggered when the user puts a finger on the screen
+   * We only care here about events with 2 fingers on them so we can control pinch to zoom
+   *
+   * @param e
+   */
+  function onDocumentTouchStart(e: React.TouchEvent<HTMLDivElement>) {
+    if (e.touches.length > 1) {
+      const startX = (e.touches[0].pageX + e.touches[1].pageX) / 2;
+      const startY = (e.touches[0].pageY + e.touches[1].pageY) / 2;
+      Object.assign(touchInfo, {
+        startX,
+        startY,
+        startDistance: Math.hypot(e.touches[1].pageX - e.touches[0].pageX, e.touches[1].pageY - e.touches[0].pageY),
+      });
+    } else {
+      Object.assign(touchInfo, {
+        startX: 0,
+        startY: 0,
+        startDistance: 0,
+      });
+    }
+  }
+
+  /**
+   * Event triggered when the user moves the finger around the screen
+   * Since we only control pinch to zoom, we need to track how the distance between the fingers changed over time
+   * Then we use that distance to calculate the relative scale and apply that scale using transforms
+   * to avoid expensive re-renders, once the user let go the fingers we do a proper rendering of the PDF document
+   *
+   * @param e
+   */
+  function onDocumentTouchMove(e: React.TouchEvent<HTMLDivElement>) {
+    if (!isTouchDevice || touchInfo.startDistance <= 0 || e.touches.length < 2) {
+      return;
+    }
+    const pinchDistance = Math.hypot(e.touches[1].pageX - e.touches[0].pageX, e.touches[1].pageY - e.touches[0].pageY);
+    const originX = touchInfo.startX + documentRef.current.scrollLeft;
+    const originY = touchInfo.startY + documentRef.current.scrollTop;
+    currentPinchScale = pinchDistance / touchInfo.startDistance;
+
+    // Adjust for min and max parameters over the absolute zoom (current zoom + pitch zoom)
+    const absScale = currentPinchScale * currentScale;
+    currentPinchScale = Math.min(Math.max(absScale, 0.2), 2.5) / currentScale;
+
+    // Here we simulate the zooming effect with transform, not perfect, but better than a re-render
+    documentRef.current.style.transform = `scale(${currentPinchScale})`;
+    documentRef.current.style.transformOrigin = `${originX}px ${originY}px`;
+  }
+
+  /**
+   * Event triggered when the user ends a touch event
+   * If all went good and we are ending a pinch to zoom event we need to queue a rendering of the PDF pages
+   * using the new zoom level
+   */
+  function onDocumentTouchEnd() {
+    if (!isTouchDevice || touchInfo.startDistance <= 0) return;
+    documentRef.current.style.transform = `none`;
+    documentRef.current.style.transformOrigin = `unset`;
+
+    const rect = documentRef.current.getBoundingClientRect();
+    const dx = touchInfo.startX - rect.left;
+    const dy = touchInfo.startY - rect.top;
+
+    // I don't like this, but we need to make sure we change the scrolling after the re-rendering with the new zoom levels
+    setTimeout(() => {
+      documentRef.current.scrollLeft += dx * (currentPinchScale - 1);
+      documentRef.current.scrollTop += dy * (currentPinchScale - 1);
+    }, 0);
+
+    Object.assign(touchInfo, {
+      startDistance: 0,
+      startX: 0,
+      startY: 0,
+    });
+    onScaleChange(currentPinchScale * currentScale);
+  }
+
+  /**
+   * Event triggered when the touch event gets cancelled
+   * In this case we need to restart our touchInfo data so other things can continue as they were
+   */
+  function onTouchCancel() {
+    if (isTouchDevice) {
+      Object.assign(touchInfo, {
+        startDistance: 0,
+        startX: 0,
+        startY: 0,
+      });
+    }
+  }
+
+  /**
    * Event triggered when a page visibility changes
    *
    * @param pageNumber
    * @param ratio
    */
-  const onPageVisibilityChanged = (pageNumber: number, ratio: number): void => {
+  const onPageVisibilityChanged = (pageNumber: number, ratio: number): boolean => {
+    // Ignore page change during pinch to zoom event
+    // This needs to be done as page changes trigger a re-rendering
+    // which conflicts with all the pinch to zoom events
+    if (isTouchDevice && window['touchInfo'].startDistance > 0) return false;
+
+    // Calculate in which page we are right now based on the scrolling position
     if (pages && pages.length) {
       pages[pageNumber - 1].ratio = ratio;
       const maxRatioPage = pages.reduce((maxIndex, item, index, array) => {
@@ -225,6 +380,7 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
     } else {
       setCurrentPage(1);
     }
+    return true;
   };
 
   /**
@@ -249,6 +405,9 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
       }
     }
   };
+  /**
+   * End of Touch events
+   */
 
   /**
    * Event triggered when the user manually changes the zoom level
@@ -281,7 +440,6 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
    * @param scale
    */
   function zoomToScale(scale: number) {
-    console.log('[-] zoomToScale', scale);
     setCurrentScale(Math.min(Math.max(scale, 0.2), 2.5));
   }
 
@@ -296,8 +454,6 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
       return;
     }
 
-    console.log('[-] zoomToPageView', viewMode);
-
     const pageElement = pageProps.ref.firstChild as HTMLDivElement;
     const pageWidth = pageProps.width || pageElement.offsetWidth;
     const pageHeight = pageProps.height || pageElement.offsetHeight;
@@ -306,21 +462,21 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
     switch (viewMode) {
       case PageViewMode.DEFAULT: {
         if (landscape) {
-          const desiredWidth = Math.round(documentRef.current.offsetWidth * 0.9);
+          const desiredWidth = Math.round(documentRef.current.offsetWidth - 32);
           zoomToScale(desiredWidth / pageWidth);
         } else {
-          const desiredWidth = Math.round(documentRef.current.offsetWidth * 0.7);
+          const desiredWidth = Math.round((documentRef.current.offsetWidth - 32) * 0.7);
           zoomToScale(desiredWidth / pageWidth);
         }
         break;
       }
       case PageViewMode.FIT_TO_WIDTH: {
-        const desiredWidth = Math.round(documentRef.current.offsetWidth * 0.95);
+        const desiredWidth = Math.round(documentRef.current.offsetWidth - 32);
         zoomToScale(desiredWidth / pageWidth);
         break;
       }
       case PageViewMode.FIT_TO_HEIGHT: {
-        const desiredHeight = Math.round(documentRef.current.offsetHeight * 0.95);
+        const desiredHeight = Math.round(documentRef.current.offsetHeight - 32);
         zoomToScale(desiredHeight / pageHeight);
         break;
       }
@@ -347,7 +503,14 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
   return (
     <PDFWorker workerUrl={workerUrl}>
       <DocumentWrapper fullscreen={fullscreen}>
-        <Document ref={documentRef}>
+        <Document
+          ref={documentRef}
+          onTouchStart={onDocumentTouchStart}
+          onTouchEnd={onDocumentTouchEnd}
+          onTouchMove={onDocumentTouchMove}
+          onTouchCancel={onTouchCancel}
+          onClick={onDocumentTouch}
+          onDoubleClick={onDocumentDoubleTouch}>
           {loading ? (
             <PageWrapper>
               <PDFViewerPage
@@ -375,7 +538,7 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
             ))
           )}
         </Document>
-        {!loading && (
+        {!loading && !isTouchDevice && (
           <PDFViewerToolbar
             labels={props.toolbarLabels}
             currentPage={currentPage}
@@ -387,6 +550,21 @@ export const PDFViewer: React.FC<PDFViewerProps> = props => {
             onScaleChange={onScaleChange}
             onViewModeChange={onViewModeChange}
             onFullscreenChange={switchFullscreenMode}
+          />
+        )}
+        {!loading && isTouchDevice && (
+          <PDFViewerTouchToolbar
+            labels={props.toolbarLabels}
+            currentPage={currentPage}
+            currentViewMode={currentViewMode}
+            numPages={pages.length}
+            currentScale={currentScale}
+            fullscreen={fullscreen}
+            onPageChange={navigateToPage}
+            onScaleChange={onScaleChange}
+            onViewModeChange={onViewModeChange}
+            onFullscreenChange={switchFullscreenMode}
+            showTouchToolbar={showTouchToolbar}
           />
         )}
       </DocumentWrapper>